# specs/002-workflow-draft-generation/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow API
  version: 0.1.0
paths:
  /workflow/sessions:
    post:
      summary: Create workflow session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                overview:
                  type: string
              required: [overview]
      responses:
        '201':
          description: Created
    get:
      summary: List workflow sessions
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /workflow/sessions/{sessionId}:
    get:
      summary: Get workflow session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    delete:
      summary: Delete workflow session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
...